+++ src/app/Components/ComponentsSamples/Pickers/ScrollPickerSamplesViewModel.cs
@@ -9,6 +9,8 @@ namespace Components.ComponentsSamples.Pickers;
 
 public class ScrollPickerSamplesViewModel
 {
+    private int? m_lastSelectedIndex;
+
     public ScrollPickerSamplesViewModel()
     {
         
@@ -47,7 +49,12 @@ public class ScrollPickerSamplesViewModel
     
     private async void OnSelectedItemChanged(int obj)
     {
-       
+        if (obj == -1)
+        {
+            await Task.Delay(1000);
+            ItemComponents[0].SetSelectedIndex(2);
+            ItemComponents[0].InvalidateData();
+        }
     }
 
     private List<int> GetWeeksInYear(int year)